requirements.txt
pillow
numpy
torch
matplotlib

# config
pyyaml